Move range ordering helper to misc
